# deploy/crm/vantriq-backend/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VantriqAI CRM API",
    "version": "9.13.0",
    "description": "The CRM behind VantriqAI: clients, invoices, payments, quotes and proposals,\ncontracts, statements, and the webhook surface the AI agents write through.\n\n## Authentication\n\nTwo ways in, and one of them is for people:\n\n- **Staff session** — `Authorization: Bearer <token>`, issued by `POST /api/auth/verify`\n  after a password and an emailed one-time code. A staff member's role decides reach.\n- **API key** — `x-api-key: <key>`, for machines. Keys are scoped.\n\n## Scopes\n\nLeast to most: `webhook` < `automation` < `staff` < `admin`. A scope opens everything\nbelow it. The admin key opens everything and is intended as break-glass only — set\n`ALLOW_API_KEY_LOGIN=false` once staff accounts exist.\n\n## Automation writes through this API, never to the database\n\nn8n and any other integration must use these endpoints rather than connecting to\nPostgres. Tenancy, quota and tax rules live here; a direct write bypasses all three."
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "crm.vantriqai.com"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "staffSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/api/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "GET /api/agents",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "POST /api/agents",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
...
    "/api/external/account": {
      "get": {
        "tags": [
          "externalApi"
        ],
        "summary": "GET /api/external/account — company, package, service status.",
        "description": "Requires the **client-token** scope or higher.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/external/agents": {
      "get": {
        "tags": [
          "externalApi"
        ],
        "summary": "GET /api/external/agents — this account's agents (channels/numbers).",
        "description": "Requires the **client-token** scope or higher.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/external/analytics": {
      "get": {
        "tags": [
          "externalApi"
        ],
        "summary": "GET /api/external/analytics?grain=day|week|month|quarter|year — the same figures as the portal's dashboard, for a customer's own BI tool. Counts only; no session ids.",
        "description": "Requires the **client-token** scope or higher.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/external/invoices": {
      "get": {
        "tags": [
          "externalApi"
        ],
        "summary": "GET /api/external/invoices — every invoice on this account.",
        "description": "Requires the **client-token** scope or higher.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/external/invoices/{id}": {
      "get": {
        "tags": [
          "externalApi"
        ],
        "summary": "GET /api/external/invoices/:id — one invoice, with its line items.",
        "description": "Requires the **client-token** scope or higher.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/external/invoices/{id}/pdf": {
      "get": {
        "tags": [
          "externalApi"
        ],
        "summary": "GET /api/external/invoices/:id/pdf — the same PDF the portal downloads.",
        "description": "Requires the **client-token** scope or higher.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/external/usage": {
      "get": {
        "tags": [
          "externalApi"
        ],
        "summary": "GET /api/external/usage[?month=YYYY-MM] — session-level usage, the same unit the invoice is billed on. Mirrors portal.js's GET /activity exactly, including what it deliberately leaves out: the raw session_id.",
        "description": "Requires the **client-token** scope or higher.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
...
    "/api/invoices": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "GET /api/invoices",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      },
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "POST /api/invoices",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/invoices/{id}": {
      "delete": {
        "tags": [
          "invoices"
        ],
        "summary": "DELETE /api/invoices/:id",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      },
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "One invoice with its printed body and its ledger.",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
...
    "/api/invoices/{id}/pdf": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "GET /api/invoices/:id/pdf — the invoice as a file you can keep. The same document the customer is emailed and the same one the portal prints, rendered from the same buildTaxInvoice() output. Before this, the CRM's Download button wrote a hand-built text file listing five fields, which is not something anyone can file or hand to an accountant.",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
...
  }
}